Add a check when building a player for OSX version
