+++ com.unity.render-pipelines.high-definition/Editor/BuildProcessors/HDRPPreprocessBuild.cs
@@ -1,3 +1,4 @@
+using UnityEngine;
 using UnityEditor.Build;
 using UnityEditor.Build.Reporting;
 using UnityEngine.Rendering;
@@ -17,7 +18,7 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
                 return;
 
             // If platform is supported all good
-            if (HDUtils.IsSupportedBuildTarget(report.summary.platform))
+            if (HDUtils.IsSupportedBuildTarget(report.summary.platform) && HDUtils.IsOperatingSystemSupported(SystemInfo.operatingSystem))
                 return;
 
             string msg = "The platform " + report.summary.platform.ToString() + " is not supported with Hight Definition Render Pipeline";
